move graviton auth api class
