+++ src/main/java/com/github/libgraviton/workerbase/gdk/GravitonAuthApi.java
@@ -1,6 +1,5 @@
-package com.github.libgraviton.workerbase;
+package com.github.libgraviton.workerbase.gdk;
 
-import com.github.libgraviton.workerbase.gdk.GravitonApi;
 import com.github.libgraviton.workerbase.gdk.api.header.HeaderBag;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
