+++ src/test/java/com/github/libgraviton/workerbase/WorkerBaseTestCase.java
@@ -1,6 +1,7 @@
 package com.github.libgraviton.workerbase;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.github.libgraviton.workerbase.gdk.GravitonAuthApi;
 import com.github.libgraviton.workerbase.gdk.api.Response;
 import com.github.libgraviton.workerbase.gdk.data.GravitonBase;
 import com.github.libgraviton.workerbase.gdk.serialization.mapper.GravitonObjectMapper;
