+++ src/test/java/com/github/libgraviton/workerbase/lib/TestFileWorker.java
@@ -2,7 +2,7 @@ package com.github.libgraviton.workerbase.lib;
 
 import com.github.libgraviton.gdk.gravitondyn.file.document.File;
 import com.github.libgraviton.workerbase.FileWorkerAbstract;
-import com.github.libgraviton.workerbase.GravitonAuthApi;
+import com.github.libgraviton.workerbase.gdk.GravitonAuthApi;
 import com.github.libgraviton.workerbase.exception.GravitonCommunicationException;
 import com.github.libgraviton.workerbase.exception.WorkerException;
 import com.github.libgraviton.workerbase.model.QueueEvent;
