+++ src/test/java/com/github/libgraviton/workerbase/lib/TestWorker.java
@@ -1,6 +1,6 @@
 package com.github.libgraviton.workerbase.lib;
 
-import com.github.libgraviton.workerbase.GravitonAuthApi;
+import com.github.libgraviton.workerbase.gdk.GravitonAuthApi;
 import com.github.libgraviton.workerbase.WorkerAbstract;
 import com.github.libgraviton.workerbase.exception.WorkerException;
 import com.github.libgraviton.workerbase.model.QueueEvent;
